cli bookmark move: deprecate multiple name arguments
There is nothing inherently wrong with moving multiple bookmarks at
once to the same revision. However, it rarely makes sense in a normal
workflow. Therefore, it is confusing to have that as a central feature
of the CLI. By only allowing a single string pattern for the bookmark to
move, the command help becomes less confusing. It is still possible to
move multiple bookmarks at once to the same revision with an appropriate
glob or regex pattern.